+++ src/main/scala/org/mmisw/orr/ont/app/UserController.scala
@@ -5,8 +5,8 @@ import com.novus.salat._
 import com.novus.salat.global._
 import com.typesafe.scalalogging.{StrictLogging => Logging}
 import org.joda.time.DateTime
-import org.mmisw.orr.ont.db
-import org.mmisw.orr.ont.{Setup, UserResult}
+import org.mmisw.orr.ont.{PasswordResetResult, db, Setup, UserResult}
+import org.mmisw.orr.ont.service.UserService
 import org.scalatra.Created
 
 import scala.util.{Failure, Success, Try}
@@ -14,6 +14,8 @@ import scala.util.{Failure, Success, Try}
 
 class UserController(implicit setup: Setup) extends BaseController
       with Logging {
+
+  val userService = new UserService
 
   createAdminIfMissing()
 
@@ -30,6 +32,139 @@ class UserController(implicit setup: Setup) extends BaseController
   get("/:userName") {
     val userName = require(params, "userName")
     getUserJson(getUser(userName))
+  }
+
+  // requests password to be reset
+  // TODO mechanism to avoid/reduce DoS attacks
+  put("/rpwr/:userName") {
+    request.getRemoteHost
+    val userName = require(params, "userName")
+    val user = getUser(userName)
+
+    new Thread(new Runnable {
+      def run() {
+        val resetRoute = s"$getMyBaseUrl/api/v0/user/auth/reset/"
+        userService.requestResetPassword(user, resetRoute)
+      }
+    }).start()
+
+    PasswordResetResult(user.userName,
+      email = Some(user.email),
+      message = Some(
+        s"""An email with password reset instructions is on its way to: ${user.email}.
+           | If you don't receive it within a few minutes, please try again later.
+         """.stripMargin)
+    )
+  }
+
+  /*
+   * Dispatches HTML form for the user to set a new password.
+   */
+  get("/auth/reset/:token") {
+    val token = require(params, "token")
+
+    contentType = formats("html")
+
+    val pwr = setup.db.pwrDAO.findOneById(token).getOrElse(
+      halt(404, s"invalid or expired token"))
+
+//    if (pwr.expiration.compareTo(DateTime.now()) > 0)
+//      error(400, "this link has expired.")
+
+    val action = request.getRequestURL.toString
+
+    <html>
+      <body>
+        <h3>
+          Set your new password:
+        </h3>
+        <form action="$action" method="post">
+          <div>
+            <label for="username">Username:</label>
+            <input type="text" id="username" name="username"/>
+          </div>
+          <div>
+            <label for="mail">E-mail:</label>
+            <input type="email" id="mail" name="email"/>
+          </div>
+          <div>
+            <label for="password">Password:</label>
+            <input type="password" id="password" name="password"/>
+          </div>
+          <div>
+            <label for="password2">Password again:</label>
+            <input type="password" id="password2" name="password2"/>
+          </div>
+          <div>
+            <input type="hidden" value="$token" name="token"/>
+          </div>
+          <div class="button">
+            <button type="submit">Submit</button>
+          </div>
+        </form>
+      </body>
+    </html>.toString()
+      .replace("$pwr", pwr.toString)
+      .replace("$action", action)
+      .replace("$token", token)
+  }
+
+  /*
+   * Form action for setting new password.
+   */
+  post("/auth/reset/:token") {
+    val token = require(params, "token")
+    val userName = require(params, "username")
+    val email = require(params, "email")
+    val password = require(params, "password")
+    val password2 = require(params, "password2")
+
+    contentType = formats("html")
+
+    val pwr = setup.db.pwrDAO.findOneById(token).getOrElse(
+      halt(404, s"invalid token"))
+
+//    if (pwr.expiration.compareTo(DateTime.now()) > 0)
+//      error(400, "sorry, this link has expired.")
+
+    val user = getUser(pwr.userName)
+
+    if (email != user.email)
+      error(400, "invalid information; click Back in your browser and try again.")
+
+    // TODO real password validation!
+    if (password.length < 8)
+      error(400, "password is too short")
+    if (password != password2)
+      error(400, "passwords don't match")
+
+    // all ok, update password:
+    val update = user.copy(password = userAuth.encryptPassword(password))
+    Try(usersDAO.update(MongoDBObject("_id" -> userName), update, false, false, WriteConcern.Safe)) match {
+      case Success(result) =>
+      case Failure(exc)    => error(500, s"update failure = $exc")
+    }
+
+    // remove token
+    Try(setup.db.pwrDAO.remove(pwr, WriteConcern.Safe)) match {
+      case Success(result) =>
+      case Failure(exc)    => error(500, s"update failure = $exc")
+    }
+
+    // TODO: update firebase.
+
+    userService.notifyPasswordHasBeenReset(user)
+
+    <html>
+      <body>
+        <h3>
+          Your password has been reset.
+        </h3>
+        Continue to <a href="$orrLink">$orrLink</a>
+        and sign in with your new password.
+      </body>
+    </html>.toString()
+      .replace("$orrLink", getMyBaseUrl)
   }
 
   /*